don’t log warnings from instUI’s DateInput or Editable
We don’t control these, those are instUI components, so it doesn’t help
To tell canvas developers that those things are doing something that is
Deprecated since we can’t do anything about it.

Yeah, if we are using a deprecated component from instUI, we want to
Know about it, but we don’t want to know if an instUI component happens
To use another instUI component in a deprecated way.

Test plan:
* some warnings should go away in the jest console logs

Change-Id: If5886b25b638c92f6d7a8e150f695ccc5ec5dc08
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/218874
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>