Add "partially enabled" state to group permissions
Closes USERS-252
Refs USERS-245
flag=granular_permissions_wiki_pages

When a set of granular permissions are group, the group
permission "rolls up" the permissions inside it. In particular
for the "enabled" setting, if some permissions are enabled and
others not, we want the group to show a "partially enabled"
state in the UI.

This is done by changing "enabled" from a Boolean to an enum,
and then converting the value into the enum for all permissions.
For "ordinary" permissions this is a simple mapping, but for
the groups there is a bit of AND/OR logic to be done.

Also converted the PermissionsButton to fully InstUI, and had
to touch a bunch of tests to make them aware of the new enabled
enum. Oh and I added tests for the new functionality in utils.js

This commit does _not_ address any of the positioning, styling,
or a11y issues with the gray checkboxes for the granular
permissions themselves... that will all get fixed up in
USERS-245.

Test plan:
* Expand the "Manage Pages" group in permissions settings
* Set some of the granular permissions for a role to a mix of
  "on" and "off"
* The button for the group permission for that role should change
  to a half-filled green circle icon
* Existing functionality for enabling/disabling group permissions
  should be the same
* API calls should still be correct
* A11y and keyboard nav for the permissions page should be right
* In particular, narrow the window so that not all of the roles
  are visible in the pane, and try keyboard tabbing: the pane
  should automatically scroll horizontally to keep the currently
  focussed button in view

Change-Id: I0bedbff9a2632f7686cba5e7ba9a0bab87fc65d6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/226803
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Steve Shepherd <sshepherd@instructure.com>