slay gradezilla
closes TALLY-565

This is entirely a search+replace of gradezilla with gradebook, merging
folders in a couple of places. There is nothing here which poses a risk
to the stability of the code. Assuming Jenkins is happy, it can be
merged.

test plan:
 * Verify Jenkins passes

Change-Id: Id35bd0fc54ff580c2e2b1513b4e34c4f76d92920
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/227231
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Keith Garner <kgarner@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>