pull image if it doesn't exist for publishing
On retriggers the image won't rebuild so the publish image
stage will fail since the image doesn't exist that we are
trying to tag and push. Added an if statement to check for
image and pull it if it doesn't exist.

flag = none

Test Plan:
Jenkins passes and retrigger passes.

Change-Id: I215d32874e643fe7d61900e070a3b75688b6efd3
closes: CCI-262
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/228562
Reviewed-by: Jacob Powell <spowell@instructure.com>
Reviewed-by: Rex Fleischer <rfleischer@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: James Butters <jbutters@instructure.com>
Product-Review: James Butters <jbutters@instructure.com>