Autofill 'type' column for CDC testing
If a model has a 'type' attribute, ActiveRecord expects
that value to be the name of a class that extends the
model. It can be safely auto-filled with the name of
any class that extends the given model.

refs PLAT-5361

Change-Id: I2251db6647551131bdacdea90b596de14b80c59d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/228699
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Evan Battaglia <ebattaglia@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>