doc: fix CalenderEventsApiController#set_course_timeable_events
events[][title] is a allowed parameter for the POST
CalenderEventsApiController#set_course_timetable_events call, however
the API documentation does not list this. Since we want to use this
parameter we would like the API documentation to be up-to-date in this
regard.

closes #1585

Test Plan:
  - Check
    /doc/api/calendar_events.html#method.calendar_events_api.set_course_timetable_events
    to see if events[][title] is listed as parameter
  - Check you can actually set the events title using this call :>.
    (relevant code at app/models/courses/timetable_event_builder.rb:162)

Change-Id: I7b2f57cc72836678e16b1aec63b591b3198d0906
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/250873
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>