add CanvasDateInput and bulk edit test table
The thrust of this commit is to implement a CanvasDateInput on top of
instui's DateInput field. The bulk editing table is just an instance
of it to test with.

closes LA-835
flag=assignment_bulk_edit

test plan:
- go to the assignment bulk editing page
- play around with the single date input component
  - selecting dates from the calendar
  - typing in dates
  - typing invalid dates
  - blank line
  - keyboard navigation
  - screen reader
- play around with your user locales and timezones and expect correct
  behavior showing the proper dates

Note that VO reads the date twice when you navigate with the keyboard. I
believe this is because it is reading the input value change and the
selected calendar date. I don't know that there's much to do about this.
The demo on instui's page does the same thing.

Change-Id: I9c66ccaa746fb91fecd20362e00a927c57f491b5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/230609
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>