Make mobile QR codes regenerate themselves in the modal
Closes USERS-439
flag=mobile_qr_login

A potential sad user path through the mobile QR code flow
would be if the code on the screen expired before they had
a chance to scan it. So let's make it automatically refresh
before it expires, and also give the user some indication
that they are life-limited (in case they try to print their
screen or something dumb like that).

Nothing was changed here in the QR code generation or
display itself, so it shouldn't be necessary to heavily QA
the functionality of the displayed codes themselves.

Test plan:
* Be sure you're set up to generate mobile login QR codes
  (see test plan for g/230318)
* Bring up a mobile QR code
* See that the modal appears with a code, and says it
  expires in 10 minutes
* Wait one minute (-ish)
* Now it should say it expires in 9 minutes
* Keep waiting or go do something else for 7-8 minutes
* Eventually the message will say "expires in a few seconds"
* Not long after that (9 minutes 45 seconds after the code
  was generated to be precise) the QR code should change to
  a new one, and the message should reset to say that the
  new code once again expires in 10 minutes
* You can verify that a new code was fetched by looking in
  the network inspector (or if you have eagle eyes by seeing
  the changes in the QR data bits themselves 😊)

Change-Id: I40ce1ab1f5e210dea12d4158bb8d11fa20fa4f42
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/230956
Reviewed-by: Keith Garner <kgarner@instructure.com>
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Keith Garner <kgarner@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>