add Observee to dashboard card
When an Observer is linked to multiple students, there is no indication
on the dashboard for which course is for which student. Now, the card
will list the names of all students the observer is linked to for that
particular course.

Test Plan:
 - Create a course with a student and linked observer
 - As the observer, the dashboard card should show the students name
 - As a user without a linked student, no observer information should be
     shown on dashboard cards.
 - An observer linked to multiple students in a course should list all
     students, the the list truncating with an ellipsis

closes KNO-366
flag=none

Change-Id: I8a5783ad9fe7d488971a98dd482632558a7357d9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/231275
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Davis Hyer <dhyer@instructure.com>
QA-Review: Davis Hyer <dhyer@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>