Adds student tour points
Builds off of previous work that has been done with Teacher tour points.
These tour points only appear for users that have a student role.

Adjusts the Courses Controller to include whether the current
user is a student in the currently viewed course as part of the
values provided to the front end.

When a user is just a student, the tour appears immediately
when visiting Canvas the first time.

When a user is both a teacher and a student, it will show the
teacher tour on every page except for the courses pages
which the user is a student in. The same applies to opening
the welcome tour from the Help tray.

Also fixes an A11y issue where VO focus isn't locked to the dialog.

Test Plan:
Make sure the Product Tours feature flag is turned on
As a student, open up canvas.
The tour should appear, allowing you to walk through the
steps.

As a student and teacher, open up canvas.
The teacher tour should appear (unless it's already appeared)
Visit the course page for the course you are enrolled in.
The student tour should appear.

Refs UXS-58

Change-Id: I0b3531b1a91ef09c402bf6240e115c02b5e8f316
(cherry picked from commit 5f4d1f1e1b22cc99594f0e8be48a5909f8f0d18a)
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/232056
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Brian Watson <bwatson@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Alex Anderson <raanderson@instructure.com>