spec: stop using error class for error_report in specs
the error report is made cause that is what canvas does to handle errors
but this method is only used in specs and when there is an error in spec
half the time it would throw a different error in here instead of the
helpful error. the other else clause added will catch when a report is
not configured correctly.

flag=none

Change-Id: I268b79a7e3ed11d080c3e10b303640620b00da64
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/232652
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
Reviewed-by: Matthew Lemon <mlemon@instructure.com>
QA-Review: Matthew Lemon <mlemon@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>