remove known unused items in gradebook actions
There is a bunch of cruft in the gradebooks controller which is not
being used anywhere. This culls some of that and moves a few items
around to be closer to where they are actually used. While these items
have been investigated for use, the related pages should be smoke tested
to confirm that nothing was overlooked.

flag = none

test plan:
 * Smoke test Default Gradebook
 * Smoke test Individual Gradebook
 * Smoke test Learning Mastery
 * Smoke test Individual Learning Mastery

Change-Id: Id19ea6b0d5763511fb495fcfb3f8503eebd50f6f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/233408
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>