Scroll expanded granular permissions so they are visible
Fixes USERS-283
flag=granular_permissions_wiki_pages

If an unexpanded group permission is at or near the bottom of
the scrolling list of permissions, clicking on the "expand"
arrow will accordion out the granulars below it but it might
not be obvious that that has happened if they are accordioned
out of sight below the scroll area.

This makes it so that the scroll area automatically scrolls
if necessary so that all the expanded granular permissions are
visible.

Note that this functionality is very hard to test in a unit
test because the dimensions of the scroll area are likely to
be zeros in the absence of the larger enclosing context. It
may be possible to test this separately in a selenium spec.

Test plan:
* Go to /accounts/default/permissions
* Find the "Manage Pages" row and position it so it is at the
  very bottom of the visible scroll area
* Click on the "expand" arrow
* The scroll area should adjust so that all the granular
  permissions are visible

Change-Id: I065beeb98e7a1eea502f946a5bebab531cc8369a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/238239
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>