Fix media comment re-init error
When you open the media recorder in the old rce or speedgrader, it
the dialog is opened from public/javascripts/media_comments.js
Before the dialog is reopened, it checks to see if it's been > 5min
since it was last opened, and if so it closes the dialog (though it's
likely already closed) and removes the mount node from the DOM. Then
it recreates the mount node, creates the dialog, and opens it.

When the above code attempts to close the dialog, setCloseOption in
app/coffeescripts/media_comments/dialog_manager.js attempts to unmount
the React component that is the recorder. Unfortunately, the component's
mount node is no longer in the DOM, and this throws an error, preventing
the dialog from opening again.

closes LA-1000
flag = none

test plan:
  - edit a Page in a course using the old rce
  - click the media toolbar button
  - close the Record/Upload media dialog
  - wait > 5 minutes
  - click the media toolbar button
  > expect the dialog to reopen

Change-Id: I86d114ac36a042de411c8ecf19c9c68ff7b37842
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/239937
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>