Use only aria-label on Files hamburger menu
The rest of the pages in Canvas use the same type of button for this
hamburger menu - the button has an aria-label but no title or text
within the button. The Files page generates the button differently.
This commit makes the button on the Files page appear consistent with
the button on all the other pages. This option was selected in
coordination with a11y team instead of option outlined in Jira for
consistency and a11y.

Fixes LS-1143
flag=none

Test plan:
 - Go to /courses/:id/files
 - Check that the blue hamburger menu in top left corner has
   aria-label and not title attribute

Change-Id: I8f0a51686b28da0a13bcd4023fd99c2f739bd92c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/241567
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>